Renamed intrcommands folder to interactions
